Working through conversion of the zonal stats functions
I got up through building the zones raster from the shapefile and this is really close to working but I ran out of time.  I need to add code to adjust the output extent to match the snap extent (right now the code is only snapping to the cellsize).  Also, the burn value will need to be set still.  I think I can set this to a field but I can't remember exactly.

The field calculator seems to be working, although I'm sure there is some expression that will fail.