Switched CDL downloader to use FTP module for download
The urllib and urllib2 modules might be a little better but until we write this to be fully Python 2/3 compatible (which is kind of annoying for urllib), using the FTP module is really easy. The CDL data in general is a pain since it is hosted on a real FTP server (so we can't use requests) and it seems like the download speed is really throttled.